+++ CHANGELOG.md
@@ -96,6 +96,7 @@
 - Replace momentjs with dayjs @robgietema
 - Add basic crud operations for content rules @robgietema
 - Add content rules actions and conditions registry @robgietema
+- Add addable actions and conditions to routes @robgietema
 
 ### Bugfix